fix: clippy warnings across examples (30+ fixes)
- Fix syntax errors: 'let x as u32' -> 'let x: u32'
- Remove unused imports (PositionSizing, HashMap, NaiveDateTime, anyhow::Result)
- Remove unnecessary .into() calls on string literals
- Remove unnecessary casts (u32 -> u32, i64 -> i64)
- Add #[allow] for intentional patterns (type_complexity, unused_imports)
- Fix unnecessary_min_or_max warnings
- Prefix unused variables with underscore
- Fix question_mark clippy (match -> ?)
- Fix redundant closures (|| expr -> expr)
- Fix borrow_deref_ref warnings
- Fix manual_range_contains
- Fix inconsistent digit grouping (3600_000 -> 3_600_000)

+++ examples/mean_reversion_test.rs
@@ -6,7 +6,7 @@
 //! Usage:
 //!   cargo run --example mean_reversion_test -- [symbol] [interval] [candles]
 
-use anyhow::Result;
+
 use krypto::{
     algo::strategies::{BollingerReversion, RsiMeanReversion},
     algo::SignalGenerator,
@@ -37,7 +37,7 @@ async fn main() -> anyhow::Result<()> {
     // Load data
     println!("Fetching data from Binance...");
     let loader = DataLoader::new(None, None);
-    let raw_df = loader.fetch_data(symbol, interval, candles).await?;
+    let raw_df = loader.fetch_data(symbol, interval, candles as u32).await?;
     println!("Loaded {} bars", raw_df.height());
 
     // Add technical indicators
